refactor: move render chunk hook to output hooks
