+++ crates/rolldown_plugin/src/plugin_driver/output_hooks.rs
@@ -1,6 +1,9 @@
+use crate::HookRenderChunkArgs;
 use crate::{HookNoopReturn, PluginDriver};
 use futures::future::join_all;
 use rolldown_common::Output;
+use rolldown_error::BuildError;
+use rolldown_sourcemap::SourceMap;
 
 impl PluginDriver {
   pub async fn render_start(&self) -> HookNoopReturn {
@@ -9,6 +12,23 @@ impl PluginDriver {
     }
     Ok(())
   }
+
+  pub async fn render_chunk(
+    &self,
+    mut args: HookRenderChunkArgs<'_>,
+  ) -> Result<(String, Vec<SourceMap>), BuildError> {
+    let mut sourcemap_chain = vec![];
+    for (plugin, ctx) in &self.plugins {
+      if let Some(r) = plugin.render_chunk(ctx, &args).await? {
+        args.code = r.code;
+        if let Some(map) = r.map {
+          sourcemap_chain.push(map);
+        }
+      }
+    }
+    Ok((args.code, sourcemap_chain))
+  }
+
   pub async fn generate_bundle(&self, bundle: &Vec<Output>, is_write: bool) -> HookNoopReturn {
     for (plugin, ctx) in &self.plugins {
       plugin.generate_bundle(ctx, bundle, is_write).await?;
