+++ packages/rolldown/src/plugin/bindingify-build-hooks.ts
@@ -6,7 +6,6 @@ import { RolldownNormalizedInputOptions } from '../options/input-options'
 import { isEmptySourcemapFiled } from '../utils'
 import path from 'path'
 import { SourceMapInputObject } from '../types/sourcemap'
-import { NormalizedOutputOptions } from '../options/output-options'
 
 export function bindingifyBuildStart(
   options: RolldownNormalizedInputOptions,
@@ -142,42 +141,6 @@ export function bindingifyLoad(
         path.resolve(directory, sourceRoot, source!),
       )
     }
-
-    return {
-      code: ret.code,
-      map: JSON.stringify(map),
-    }
-  }
-}
-
-export function bindingifyRenderChunk(
-  outputOptions: NormalizedOutputOptions,
-  hook?: Plugin['renderChunk'],
-): BindingPluginOptions['renderChunk'] {
-  if (!hook) {
-    return undefined
-  }
-  const [handler, _optionsIgnoredSofar] = normalizeHook(hook)
-
-  return async (code, chunk) => {
-    const ret = await handler.call(null, code, chunk, outputOptions)
-
-    if (ret == null) {
-      return
-    }
-
-    if (typeof ret === 'string') {
-      return { code: ret }
-    }
-
-    if (!ret.map) {
-      return { code: ret.code }
-    }
-
-    let map =
-      typeof ret.map === 'object'
-        ? ret.map
-        : (JSON.parse(ret.map) as SourceMapInputObject)
 
     return {
       code: ret.code,
