+++ packages/rolldown/src/plugin/bindingify-output-hooks.ts
@@ -18,6 +18,38 @@ export function bindingifyRenderStart(
     handler.call(null, outputOptions, options)
   }
 }
+
+export function bindingifyRenderChunk(
+  outputOptions: NormalizedOutputOptions,
+  hook?: Plugin['renderChunk'],
+): BindingPluginOptions['renderChunk'] {
+  if (!hook) {
+    return undefined
+  }
+  const [handler, _optionsIgnoredSofar] = normalizeHook(hook)
+
+  return async (code, chunk) => {
+    const ret = await handler.call(null, code, chunk, outputOptions)
+
+    if (ret == null) {
+      return
+    }
+
+    if (typeof ret === 'string') {
+      return { code: ret }
+    }
+
+    if (!ret.map) {
+      return { code: ret.code }
+    }
+
+    return {
+      code: ret.code,
+      map: typeof ret.map === 'object' ? JSON.stringify(ret.map) : ret.map,
+    }
+  }
+}
+
 export function bindingifyGenerateBundle(
   hook?: Plugin['generateBundle'],
 ): BindingPluginOptions['generateBundle'] {
